+++ src/data/hanzi-dictionary.ts
@@ -1,20 +1,6 @@
-// ADR-0003: Hanzi dictionary structure
-export type Tone = 1 | 2 | 3 | 4;
+import type { HanziEntry, HanziMeaning } from '@/types/hanzi';
 
-export interface HanziMeaning {
-  id: string; // "說-yuè-1"
-  onyomi: string; // "エツ"
-  pinyin: string; // "yuè"
-  tone: Tone; // 4
-  meaning_ja: string; // "喜ぶ"
-  is_default: boolean;
-}
-
-export interface HanziEntry {
-  id: string; // "說"
-  meanings: HanziMeaning[];
-  is_common: boolean;
-}
+export type { HanziEntry, HanziMeaning };
 
 export const hanziDictionary: HanziEntry[] = [
   {
